+++ changelog.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- Remove unused journal logging (#431)
+
 ## v1.0.1302 (2018-07-26)
 
 - Authentication added for docker build (#449)